Load, check and plot input data for use in breakpoint detection

Load, check and plot input data for use in breakpoint detection

Usage

check_data(dat, sex = FALSE, showPlot = TRUE)

Arguments

dat

data.frame with columns year, age, length, sex (optional)

sex

logical. does your data frame include a sex column?

showPlot

logical. do you want to visualize your input data?

Value

plots