Wrapper function to to fit gams & evaluate first derivatives

Wrapper function to to fit gams & evaluate first derivatives

Usage

get_Breaks(dat, ages_to_use = c(5, 10), sex = FALSE, axes = 0, showPlot = TRUE)

Arguments

dat

data.frame with columns year, age, length, lat, long, sex (optional)

ages_to_use

optional. vector of age(s) to evaluate for breakpoints.

sex

logical.

axes

do you want to evaluate axes for space only (0, default), time only (1), or both simultaneously (2)?

showPlot

logical. do you want to see the detected break(s) on a map? Applicable only for axes options 0 or 2.

Value

matrix of detected breakpoints and plots

Examples

 if (FALSE)  data(simulated_data) ;get_Breaks(dat = simulated_data, ages_to_use = c(15:20), axes = 0)
#> Error in dat %>% group_by(age) %>% group_split(): could not find function "%>%"
 # \dontrun{}