Re-fit growth data at putative breaks and return estimates for validation

Re-fit growth data at putative breaks and return estimates for validation

Usage

refit_Growth(dat = simulated_data, breakpoints, selex = FALSE, showPlot = TRUE)

Arguments

dat

data.frame with columns year, age, length, lat, long, sex (optional)

breakpoints

data.frame with columns year and/or lat and long. can be output of get_Breaks().

showPlot

logical. do you want to see plots of the fitted curves?

Value

Von B growth parameters at input breakpoints; plots with uncertainty of growth curves:

$split_tables

list of tables of input data split by strata specified in breakpoints

$fits_df

input data, estimates and associated standard errors as single dataframe

$pars_df

Parameter estimates and associated standard errors

$fits_plot

input observations and fitted growth curves, by strata

$pars_plot

Parameter estimates and associated standard errors; red lines indicate statistically significant differences